backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0
reportlab==4.2.5
